No Response as return in the service methods
The Response class is no longer used as return in the service methods.
This class was introduced to customize the return status codes but, as
currently this is done using WebApplicationException, the regular type
of the objects returned can be now used.

This update also includes a simplification of the UserAccessChecker
class' "doIfPrivileged" method.