Raw enum values in GeneDrugGroupInfo
The type of the enum attributes in GeneDrugGroupInfo has been changed to
the enum type, avoiding the conversion of the enums into Strings. For
cancer, now an array of enums is stored instead of a comma-separated
values string.

The GeneDrugGroupInfo is returned by the GeneDrugService, therefore, the
REST API has changed.