Adds the gene-gene interactions as a relation in the database
The gene-gene interactions are now stored in the database. Until now,
these relations were calculated from the gene-protein and
protein-protein interactions already stored in the database. As these
operation implied several table joins and the gene-gene interactions are
commonly used, it is interesting to have these relations already
calculated in the database.