Adds support for VEP process interruption
If the VEP process returns a 130 exit-code, which means that the process was
interrupted. This occurs, for example, when the tomcat server stops while VEP
child processes are running. We want that this process is marked as "Interrupted"
but not finished (overall progress != 1.0), so this computation can be resumed
afterwards, for example, when the servers starts again.