Adds support for gene-drug queries by drug
This commits adds a new query methods that uses the standard drug names
as the filtering criteria. In addition, a new method to get the standard
name of the drugs in the database has been added.

Due to the inclusion of genes and drugs not present in the gene-drugs
relations, this commit also moves the gene and drug names list methods
to the GeneDrugService, so that the values returned are only values
present in the gene-drugs relations.