Changes the drug and gene score calculation for drug queries
When an user performs a gene-drug query by drug, he's mainly interested
in the drug information and not in finding drug candidates for patients
with certain genes affected. Therefore, the gene-drugs and the gene-drug
group returned in a drug query contain the drug and genes scores that
are stored in the database, without any transformation.