Renames the packages to replace pandrugsdb by pandrugs
This is the first commit of two to rename the application from PanDrugsDB to
PanDrugs. In this step, the package folders have been changed from pandrugsdb
to pandrugs.

The application renaming is done in two steps so that Git can track the file
and folder renaming, as it can't track renaming and modifying a file at the
same time.