Adds Entrez id to the genes
This commit adds the Entrez ids to the genes. The gene symbol is still
the primary id of each gene, but now all the Entrez ids of each gene are
stored and accessible for each gene.

Regarding to the REST API, on the gene-drug queries now for each direct
or indirect gene both the gene symbol and the Entrez ids are returned
instead of returning only the gene symbol.