Fixes the drug status info message generation
The drug status info text generation was not updated when the entrez id and
pathway information was added to the direct and indirect gene information. This
caused that instead of inserting the gene symbol in the status info text, the
object's "toString()" were inserted.

This commit fixes this problem, using the gene symbol again.