Adds gene drug warnings
Some gene-drug relations may have some issues that should be notified to
the user. This commit adds a new GeneDrugWarning entity that contains
this information.

We decided to create a new entity instead of including this information
in the GeneDrug entity as it also affects to indirect relations (i.e.
not only to direct gene-drug relations). The warnings are added by the
GeneDrugController after the gene-drug groups are created.