Fixes gene-drug sources
The source of each gene-drug used was the source of the drugs. As the
source of the drugs may come from different gene-drug relations, this
was incorrect (i.e. each gene-drug was associated to the sources of
their drug in every gene-drug relation).

This commit fixes this error by introducing a new many-to-many
relationship between GeneDrug and DrugSource.