Adds support for indirect resistances
Some genes may have an indirect resistance to drugs caused for the
overactivation of another genes. This may cause a gene sensible to a
drug to act as if it have a resistance.

This commit adds support to include this kind of indirect resistances by
introducing the IndirectResistance entity. The indirect resistance
information is used by the GeneDrugController, that adds it to the
corresponding GeneDrugGroups.