Fixes bugs in GeneDrugInfo
This commit fixes two bugs in the GeneDrugInfo.

The first bug is that the drugStatusInfo messages where using the
standard drug name instead of the show drug name. To provide full
information to the clients, a new field showDrugName has been added with
the latter.

The second bug is that, when joining the gene names, if the gene-drug
contains more that two genes, the last one was repeated in the string
generated.