Modifies the interaction filters
Gene-drug queries currently allow filtering the results based on two
gene features: target/marker and direct/indirect. However, the
interaction type of the gene-drug relations are classified as direct
target (gene is direct and target), biomarker (gene is direct and
marker) and pathway member (gene is indirect and target).

This commit changes the gene-drug filters from target/marker and
direct/indirect to direct target, biomarker and pathway member. This new
filtering options allow a fine grained filtering of the results and is
more coherent with the user's needs.