Adds support for multiple gene/drug warnings
The GeneDrugWarning entity allows adding warnings for specific gene-drug
relations. Until now, the warnings were specific for a gene-drug pair,
and it was not possible to specify warning for a single gene or for a
single drug (this is, for all the gene-drug relations with an specific
gene or drug).

This commit modifies PanDrugs to support specific gene or drug warnings
by supporting * values in gene or drug column of the gene_drug_warning
table.