Modifies indirect resistances management
Currently, when an indirect resistance was present for a gene-drug
relation its DScore was converted to a negative value and the
sensitivity was changed to "Resistance".

This behavior has been changed, and now the indirect resistances are
treated as a type of warning, generating the corresponding warning
message. Neither the DScore or the sensitivity of the gene-drug are
changed now.