Fixes a regression problem with variant analysis files
In commit a55b7951, the usage of Apache Commons' FileUtils.readLines
was replaced with the Java NIO's Files.lines method. However, the
latter expect files to be in UTF-8 while the former expect them to
be in ISO-8859-1. This caused variant results to not load when some
special characters are present in a file.

This bug fixes this error by specifying that files are in ISO-8859-1.