Modifies entity field sizes to support UTF8 db encoding
The field drug.standard_drug_name was too big to be used as primary
key in a MySQL database using UTF8 encoding.

This commit reduces the size of this field and the foreign key fields
pointing to this field to allow the use of UTF8 encoding.

In addition, the size of the gene_drug_to_drug_source.alteration
field was increased in order to support new incoming data.

In order to ease working with a Pandrugs database, SQL scripts for
database creation and table dropping have been added to the project.