Remove NOT_ALTERED labels set to CNV/expression data
So far, CNV and/or expression genes were marked as NOT_ALTERED when adding
the SNV data to the MultiOmicsAnalysisQueryData. This caused that all
genes in the "SNV data" (snvData map) to be added to the query genes list
in the "getQueryGenes". Due to this, all expression geneswere added to the
query, although only the highly overexpressed ones must be added. This
commit fixes this problem and removes the NOT_ALTERED labels. This way, the
queries are only performed with CNV, altered SNV, and highly overexpressed
genes.