Adds support for gene dependencies to the main query logic
These are considered a new type of indirect gene, along with metabolic
pathway members.

Therefore, queries now include an option to indicate whether or not to
include gene dependencies and query results include these new gene-drug
relationships based on gene dependencies. In order to facilitate the
identification of the type of gene-drug relationship, an attribute that
specifies the type of relationship is included in the entities returned
in the REST layer.

In future commits, queries with CNV and expression data should be
modified to take into account the gene dependency data.