Fixes unrequested indirect interaction type gene-drug problem
When a request was made for a single indirect interaction type (i.e.,
pathway member or gene dependency) and the results included gene-drugs
with query genes in both indirect types, they were included in the
results with both interaction types and not just with the requested
type.

This commit solves this problem by explicitly specifying to the
GeneDrugGroup the type of indirect relations requested.